pyproject.toml
[project]
name = "extreme-temperatures"
version = "0.1.0"
description = "Extreme temperature analysis tool - fetch and analyze historical weather data from NOAA GSOD and GHCN Daily"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery",
    "pandas",
    "db-dtypes",
    "pyarrow",
    "requests",
    "marimo",
    "plotly",
    "scipy",
    "pydantic-ai-slim>=1.56.0",
    "google-genai>=1.62.0",
]

[project.scripts]
extreme-temperatures = "weather_fetcher.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/weather_fetcher"]

[tool.uv.workspace]
members = ["backend"]
